add "bbox" to __geo_interface__ output
 http://www.geojson.org/geojson-spec.html#bounding-boxes

Most geometry and feature objects already had a bound method to determine the bounding box coordinates but the coordinates were not being added to the __geo_interface__ output under the "bbox" member.

It is unclear what is the proper way to handle the "bbox" member for points.  Can/should a point have a bbox?  Since the bbox member is optional I am omitting it from point objects.